refactored cartridge logic
